Split parser.rs into seperate files
