+++ src/builtin_parser/parser.rs
@@ -1,9 +1,15 @@
 //! Generates an abstract syntax tree from a list of tokens.
 
-use kinded::Kinded;
+mod ast;
+mod error;
+#[cfg(test)]
+mod tests;
+
+pub use ast::*;
+pub use error::*;
+
 use logos::Span;
 use std::collections::HashMap;
-use std::fmt::Display;
 use std::num::IntErrorKind;
 
 use crate::builtin_parser::number::{
@@ -14,11 +20,6 @@ use super::lexer::{FailedToLexCharacter, Token, TokenStream};
 use super::number::Number;
 use super::runner::environment::Function;
 use super::{Diagnostic, Environment, SpanExtension, Spanned};
-
-/// An [Abstract Syntax Tree](https://en.wikipedia.org/wiki/Abstract_syntax_tree).
-///
-/// This type represents a list of expressions, which is what makes up a command.
-pub type Ast = Vec<Spanned<Expression>>;
 
 macro_rules! expect {
     ($tokens:ident, $($token:tt)+) => {
@@ -36,241 +37,6 @@ macro_rules! expect {
             None => return Err($tokens.span().diagnose(ParseError::ExpectedMoreTokens)),
         }
     };
-}
-
-/// A type that represents an expression.
-#[derive(Debug, Clone, Kinded)]
-pub enum Expression {
-    // Primitives
-    None,
-    Boolean(bool),
-    Number(Number),
-    Variable(String),
-    String(String),
-    Borrow(Box<Spanned<Expression>>),
-    Dereference(Box<Spanned<Expression>>),
-    Object(HashMap<String, Spanned<Expression>>),
-    StructObject {
-        name: String,
-        map: HashMap<String, Spanned<Expression>>,
-    },
-    Tuple(Vec<Spanned<Expression>>),
-    StructTuple {
-        name: String,
-        tuple: Vec<Spanned<Expression>>,
-    },
-
-    // Expressions
-    BinaryOp {
-        left: Box<Spanned<Expression>>,
-        operator: BinaryOperator,
-        right: Box<Spanned<Expression>>,
-    },
-    UnaryOp {
-        operator: UnaryOperator,
-        operand: Box<Spanned<Expression>>,
-    },
-    Member {
-        left: Box<Spanned<Expression>>,
-        right: Spanned<Access>,
-    },
-
-    // Statement-like
-    VarAssign {
-        name: Box<Spanned<Expression>>,
-        value: Box<Spanned<Expression>>,
-    },
-    Function {
-        name: String,
-        arguments: Vec<Spanned<Expression>>,
-    },
-    ForLoop {
-        index_name: String,
-        loop_count: u64,
-        block: Ast,
-    },
-}
-
-impl ExpressionKind {
-    pub const fn as_natural(&self) -> &'static str {
-        match self {
-            ExpressionKind::None => "nothing",
-            ExpressionKind::Boolean => "a boolean",
-            ExpressionKind::Number => "a number",
-            ExpressionKind::Variable => "a variable name",
-            ExpressionKind::String => "a string",
-            ExpressionKind::Borrow => "a borrow",
-            ExpressionKind::Dereference => "a dereference",
-            ExpressionKind::Object => "an object",
-            ExpressionKind::StructObject => "a struct object",
-            ExpressionKind::Tuple => "a tuple",
-            ExpressionKind::StructTuple => "a struct tuple",
-
-            ExpressionKind::BinaryOp => "a binary operation",
-            ExpressionKind::UnaryOp => "a unary operation",
-            ExpressionKind::Member => "a member expression",
-            ExpressionKind::VarAssign => "a variable assignment",
-            ExpressionKind::Function => "a function call",
-            ExpressionKind::ForLoop => "a for loop",
-        }
-    }
-}
-
-/// A singular element access within a [`Expression::Member`].
-///
-/// Based on `bevy_reflect`'s `Access`.
-#[derive(Debug, Clone, Kinded)]
-pub enum Access {
-    /// A name-based field access on a struct.
-    Field(String),
-    /// An index-based access on a tuple.
-    TupleIndex(usize),
-    // /// An index-based access on a list.
-    // ListIndex(usize),
-}
-impl AccessKind {
-    /// Returns the kind of [`Access`] as a [string slice](str) with an `a` or `an` prepended to it.
-    /// Used for more natural sounding error messages.
-    pub const fn as_natural(self) -> &'static str {
-        match self {
-            AccessKind::Field => "a field",
-            AccessKind::TupleIndex => "a tuple",
-        }
-    }
-}
-
-/// Get the access if its of a certain type, if not, return a [`EvalError`](super::runner::error::EvalError).
-///
-/// For examples, take a look at existing uses in the code.
-macro_rules! access_unwrap {
-    ($expected:literal, $($variant:ident($variant_inner:ident))|+ = $val:expr => $block:block) => {{
-        let Spanned { span, value } = $val;
-        if let $(Access::$variant($variant_inner))|+ = value $block else {
-            use $crate::builtin_parser::parser::AccessKind;
-            use $crate::builtin_parser::runner::error::EvalError;
-
-            Err(span.diagnose(
-                EvalError::IncorrectAccessOperation {
-                    expected_access: &[$(AccessKind::$variant),+],
-                    expected_type: $expected,
-                    got: value.kind(),
-                },
-            ))?
-        }
-    }};
-}
-pub(crate) use access_unwrap;
-
-#[derive(Debug, Clone)]
-pub enum BinaryOperator {
-    Add,
-    Sub,
-    Mul,
-    Div,
-    Mod,
-
-    And,
-    Xor,
-    Or,
-}
-impl Display for BinaryOperator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            BinaryOperator::Add => write!(f, "+"),
-            BinaryOperator::Sub => write!(f, "-"),
-            BinaryOperator::Mul => write!(f, "*"),
-            BinaryOperator::Div => write!(f, "/"),
-            BinaryOperator::Mod => write!(f, "%"),
-            BinaryOperator::And => write!(f, "&"),
-            BinaryOperator::Xor => write!(f, "^"),
-            BinaryOperator::Or => write!(f, "|"),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum UnaryOperator {
-    /// `-x`
-    Minus,
-    /// `!true`
-    Not,
-}
-impl Display for UnaryOperator {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            UnaryOperator::Minus => write!(f, "-"),
-            UnaryOperator::Not => write!(f, "!"),
-        }
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum ParseError {
-    #[error("unknown token character: {0}")]
-    FailedToLexCharacters(String),
-
-    #[error("expected more tokens, got nothing.")]
-    ExpectedMoreTokens,
-
-    #[error("expected token {expected:?}, got token {got:?} instead.")]
-    ExpectedTokenButGot { expected: Token, got: Token },
-
-    #[error(
-        "expected a semicolon or endline after a complete statement, but got more tokens than expected."
-    )]
-    ExpectedEndline(Token),
-
-    #[error("expected a literal token, got {0:?} which is not a valid literal.")]
-    ExpectedLiteral(Token),
-
-    #[error(
-        "invalid suffix `{0}` for number literal. the suffix must be one of the numeric types (`u32`, `isize`, `f32`, etc.)"
-    )]
-    InvalidSuffixForNumber(String),
-
-    #[error(
-        "invalid suffix `{0}` for float literal. the suffix must be one of the float types (`f32`, `f64`)"
-    )]
-    InvalidSuffixForFloat(String),
-
-    #[error(r#""{0}" is an invalid suffix for a float. the valid suffixes are "f32" and "f64"."#)]
-    IntegerSuffixOnFloat(String),
-
-    #[error("{number} cannot be represented as {number_kind:#} as it is too small.")]
-    NegativeIntOverflow {
-        number: String,
-        number_kind: crate::builtin_parser::number::NumberKind,
-    },
-
-    #[error("{number} cannot be represented as {number_kind:#} as it is too large.")]
-    PositiveIntOverflow {
-        number: String,
-        number_kind: crate::builtin_parser::number::NumberKind,
-    },
-
-    #[error(
-        "expected a continuation to the object declaration (such as a comma or a closing bracket), but got {0:?} instead."
-    )]
-    ExpectedObjectContinuation(Option<Result<Token, FailedToLexCharacter>>),
-
-    #[error(
-        "expected an identifier or integer when accessing member of variable, got {got:?} instead."
-    )]
-    ExpectedIndexer { got: Token },
-
-    #[error("{feature} are not yet supported. {}", format_issue(*issue))]
-    UnsupportedFeature { feature: &'static str, issue: u8 },
-
-    #[error("unexpected closing delimiter: `{delimiter}`")]
-    MismatchedDelimiter { delimiter: char },
-}
-
-fn format_issue(issue: u8) -> String {
-    if issue != 0 {
-        format!("see bevy_dev_console issue #{issue}")
-    } else {
-        String::new()
-    }
 }
 
 const FLOAT_PARSE_EXPECT_REASON: &str =
@@ -862,105 +628,4 @@ fn parse_object(
     }
     expect!(tokens, Token::RightBracket);
     Ok(map)
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::builtin_parser::Spanned;
-    use crate::builtin_parser::parser::Expression;
-    use logos::Span;
-
-    use super::super::Environment;
-    use super::super::lexer::TokenStream;
-    use super::Expression::*;
-    use super::parse;
-    use crate::builtin_parser::Integer::*;
-    use crate::builtin_parser::Number;
-    use crate::builtin_parser::SignedInteger::*;
-    use crate::builtin_parser::parser::BinaryOperator::*;
-    use std::assert_matches;
-
-    fn setup(src: &str) -> std::vec::IntoIter<Spanned<Expression>> {
-        let mut lexer = TokenStream::new(src);
-        let environment = Environment::default();
-
-        let ast = parse(&mut lexer, &environment).unwrap();
-
-        ast.into_iter()
-    }
-
-    #[test]
-    fn var_assign() {
-        let mut stmts = setup("x = 1 + 2 - 30 + y");
-
-        assert_matches!(
-            stmts.next(),
-            Some(Spanned {
-                span: Span { start: 0, end: 18 },
-                value:
-                    VarAssign {
-                        name:
-                            box Spanned {
-                                span: Span { start: 0, end: 1 },
-                                value: Variable(_),
-                            },
-                        value:
-                            box Spanned {
-                                span: Span { start: 4, end: 18 },
-                                value:
-                                    BinaryOp {
-                                        left:
-                                            box Spanned {
-                                                span: Span { start: 4, end: 14 },
-                                                value:
-                                                    BinaryOp {
-                                                        left:
-                                                            box Spanned {
-                                                                span: Span { start: 4, end: 9 },
-                                                                value:
-                                                                    BinaryOp {
-                                                                        left:
-                                                                            box Spanned {
-                                                                                span:
-                                                                                    Span {
-                                                                                        start: 4,
-                                                                                        end: 5,
-                                                                                    },
-                                                                                value:
-                                                                                    Number(Number::Integer(Signed(Unspecified(1)))),
-                                                                            },
-                                                                        operator: Add,
-                                                                        right:
-                                                                            box Spanned {
-                                                                                span:
-                                                                                    Span {
-                                                                                        start: 8,
-                                                                                        end: 9,
-                                                                                    },
-                                                                                value:
-                                                                                    Number(Number::Integer(Signed(Unspecified(2)))),
-                                                                            },
-                                                                    },
-                                                            },
-                                                        operator: Sub,
-                                                        right:
-                                                            box Spanned {
-                                                                span: Span { start: 12, end: 14 },
-                                                                value: Number(Number::Integer(Signed(Unspecified(30)))),
-                                                            },
-                                                    },
-                                            },
-                                        operator: Add,
-                                        right:
-                                            box Spanned {
-                                                span: Span { start: 17, end: 18 },
-                                                value: Variable(_),
-                                            },
-                                    },
-                            },
-                    },
-            })
-        );
-        assert!(stmts.next().is_none());
-    }
 }
